Simplify inclusion of OMX service based on bitness
OMX is only available for 32 bits. To simplify the handling of OMX in
64-bits only vs. dual ABI targets, the OMX HAL is extracted into a
manifest fragment, which is ignored if the build does not support
32-bit apps.

Now that libffmpeg_omx is only compiled for 32 bits target, it will only
be compiled in when needed.

OMX and codec2 services are part of the core system, so they do not need
to be explicitly specified in PRODUCT_PACKAGES.