Refactor the audio initialization stage
- If /data/vendor/asound.state is found, initialize with alsactl
and then restore asound.state
- If /data/vendor/asound.state is not found, we will start initialize
and save the alsa state with these steps:

1. Initialize with alsactl
2. Restore alsa state that are already available at /system/etc/alsa
or /data/vendor/alsa
If no state found or the state are not the one match the card name,
turn on everything using amixer (the value are based on PCH cards)
2.5. Set HiFi verb & enable devices using alsaucm for hardware that
need this. We put this before the amixer stage is because we want
the audio to be in the default volume instead of set everything to
100 and might kill a speaker or two
3. When everything done, store the state to /data/vendor/asound.state

With this new audio initialization, we can now remove all the alsa
state in /system/etc/alsa. We still keep the restore part for this
though in case someone really want to ultilize it again.

Signed-off-by: hmtheboy154 <buingoc67@gmail.com>